fix cache return
