+++ test/middleware/cache.js
@@ -84,7 +84,6 @@ describe('cache', () => {
         const request = supertest(server);
 
         const response1 = await request.get('/test/cache');
-        await wait(3000); // Wait for cache write
         const response2 = await request.get('/test/cache');
 
         const parsed1 = await parser.parseString(response1.text);
@@ -140,7 +139,6 @@ describe('cache', () => {
         const request = supertest(server);
 
         const response1 = await request.get('/test/cache');
-        await wait(3000); // Wait for cache write
         const response2 = await request.get('/test/cache');
 
         const parsed1 = await parser.parseString(response1.text);
